pyproject.toml
[project]
authors = [{name = "heyjiacheng", email = "your_xujiacheng1016@hotmail.com"}]
name = "POGS"
requires-python = ">= 3.10"
version = "0.1.0"
dependencies=[
    "Pillow",
    "jaxtyping",
    "rich",
    "open-clip-torch",
    "numpy==1.26.4",
    "torchtyping",
    "autolab_core",
    "moviepy",
    "kornia",
    "iopath",
    "transformers==4.44.0",
    "typeguard>=4.0.0",
    "awscli",
    "pyrealsense2"
, "pycocotools>=2.0.10,<3"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["pogs*"]

[project.entry-points.'nerfstudio.method_configs']
pogs = 'pogs.pogs_config:pogs_method'

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[dependency-groups]
rekep = ["pyrealsense2>=2.56.5.9235,<3", "kmeans-pytorch>=0.3,<0.4", "opencv-python", "openai>=1.97.1,<2", "numba>=0.61.2,<0.62", "rich>=14.1.0,<15", "parse>=1.20.2,<2", "imageio>=2.37.0,<3", "dds-cloudapi-sdk==0.4.1", "ur-rtde>=1.6.1,<2"]

[tool.pixi.workspace]
channels = ["nvidia/label/cuda-11.8.0", "conda-forge"]
platforms = ["linux-64"]

[tool.pixi.pypi-dependencies]
torch       = { version = "==2.6.0", index = "https://download.pytorch.org/whl/cu118" }
torchvision = { version = "==0.21.0", index = "https://download.pytorch.org/whl/cu118" }
pogs = { path = ".", editable = true }

[tool.pixi.tasks]
build = """
pip install ninja &&
pip install git+https://github.com/NVlabs/tiny-cuda-nn/#subdirectory=bindings/torch &&
pip install jaxtyping rich &&
pip install gsplat --index-url https://docs.gsplat.studio/whl/pt20cu118 &&
pip install warp-lang &&
pip install --extra-index-url=https://pypi.nvidia.com "cudf-cu11==25.4.*" "cuml-cu11==25.4.*" &&
python -m pip install -e . &&
pip install timm==0.6.7 &&
python -m pip install src/pogs/dependencies/nerfstudio/ &&
pip install fast_simplification==0.1.7 &&
pip install git+https://github.com/facebookresearch/detectron2.git &&
pip install git+https://github.com/openai/CLIP.git &&
ns-install-cli &&
pip install ur_rtde==1.4.2 cowsay opt-einsum pyvista autolab-core &&
pip install -e ./src/pogs/dependencies/ur5py &&
cd ./src/pogs/dependencies/raftstereo &&
bash download_models.sh &&
pip install -e .
"""

rekep = """
pixi run python scripts/move_to_home.py &&
pixi run -e rekep python scripts/auto_pipeline.py
"""

[tool.pixi.dependencies]
python = "3.10.*"
pip = ">=25.1.1,<26"
cuda-toolkit = ">=11.8.0,<12"
numpy = "1.26.4.*"

[tool.pixi.feature.rekep.dependencies]
supervision = ">=0.25.1,<0.26"
scikit-learn = ">=1.7.1,<2"
open3d = ">=0.19.0,<0.20"

[tool.pixi.environments]
rekep = ["rekep"]
